pyproject.toml
[project]
name = "kyf-furia-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "cassmatheuss",email = "cassmatheuss@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flask-cors (>=5.0.1,<6.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "pymongo (>=4.12.0,<5.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "langchain-openai (>=0.3.14,<0.4.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "face-recognition (>=1.3.0,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models",
    "setuptools (>=80.0.0,<81.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
